feat: implement adapter_ha pooled connector tests and refactor related logic
